initdata: validate initdata by unmarshal it - Validate the initdata passed in both from configmap and annotation Fixes: #1985
